+++ rollup.config.js
@@ -7,7 +7,9 @@ import json from 'rollup-plugin-json'
 import flow from 'rollup-plugin-flow'
 import { terser } from 'rollup-plugin-terser'
 import sourceMaps from 'rollup-plugin-sourcemaps'
-import ignore from 'rollup-plugin-ignore'
+
+// rollup-plugin-ignore stopped working, so we'll just remove the import line 😐
+const streamIgnore = { "import stream from 'stream';": "'';" }
 
 const cjs = {
   exports: 'named',
@@ -69,9 +71,9 @@ const umdBaseConfig = Object.assign({}, configBase, {
   external: Object.keys(globals),
   plugins: configBase.plugins.concat(
     replace({
+      ...streamIgnore,
       __SERVER__: JSON.stringify(false),
-    }),
-    ignore(['stream'])
+    })
   ),
 })
 
@@ -131,9 +133,9 @@ const browserConfig = Object.assign({}, configBase, {
   ],
   plugins: configBase.plugins.concat(
     replace({
+      ...streamIgnore,
       __SERVER__: JSON.stringify(false),
-    }),
-    ignore(['stream'])
+    })
   ),
 })
 
@@ -202,9 +204,9 @@ const noParserBrowserConfig = Object.assign({}, configBase, {
   ],
   plugins: configBase.plugins.concat(
     replace({
+      ...streamIgnore,
       __SERVER__: JSON.stringify(false),
-    }),
-    ignore(['stream'])
+    })
   ),
 })